# lib/api-zod/src/generated/types/chatbotStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NeuroDesk AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatbotStats {
  totalConversations: number;
  totalMessages: number;
  averageMessagesPerConversation: number;
  /** @nullable */
  trainingScore?: number | null;
  sourceCount: number;
}